+++ app/src/main/java/org/kiwix/kiwixmobile/di/components/KiwixActivityComponent.kt
@@ -22,6 +22,7 @@ import dagger.BindsInstance
 import dagger.Subcomponent
 import org.kiwix.kiwixmobile.core.di.ActivityScope
 import org.kiwix.kiwixmobile.core.di.components.CoreActivityComponent
+import org.kiwix.kiwixmobile.core.webserver.ZimHostModule
 import org.kiwix.kiwixmobile.di.modules.KiwixActivityModule
 import org.kiwix.kiwixmobile.intro.IntroFragment
 import org.kiwix.kiwixmobile.intro.IntroModule
@@ -32,8 +33,6 @@ import org.kiwix.kiwixmobile.nav.destination.library.LocalLibraryFragment
 import org.kiwix.kiwixmobile.nav.destination.library.OnlineLibraryFragment
 import org.kiwix.kiwixmobile.nav.destination.reader.KiwixReaderFragment
 import org.kiwix.kiwixmobile.settings.KiwixSettingsFragment
-import org.kiwix.kiwixmobile.webserver.ZimHostFragment
-import org.kiwix.kiwixmobile.webserver.ZimHostModule
 import org.kiwix.kiwixmobile.zimManager.fileselectView.effects.DeleteFiles
 
 @ActivityScope
@@ -50,7 +49,6 @@ interface KiwixActivityComponent : CoreActivityComponent {
   fun inject(deleteFiles: DeleteFiles)
   fun inject(localFileTransferFragment: LocalFileTransferFragment)
   fun inject(languageFragment: LanguageFragment)
-  fun inject(zimHostFragment: ZimHostFragment)
   fun inject(kiwixSettingsFragment: KiwixSettingsFragment)
   fun inject(introActivity: IntroFragment)
   fun inject(kiwixMainActivity: KiwixMainActivity)
