+++ core/src/main/java/org/kiwix/kiwixmobile/core/di/modules/ApplicationModule.kt
@@ -22,6 +22,7 @@ import android.app.DownloadManager
 import android.app.NotificationManager
 import android.content.Context
 import android.net.ConnectivityManager
+import android.net.wifi.WifiManager
 import android.os.storage.StorageManager
 import dagger.Module
 import dagger.Provides
@@ -96,4 +97,9 @@ class ApplicationModule {
   @Singleton
   fun provideConnectivityManager(context: Context): ConnectivityManager =
     context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+
+  @Provides
+  @Singleton
+  fun provideWifiManager(context: Context): WifiManager =
+    context.applicationContext.getSystemService(Context.WIFI_SERVICE) as WifiManager
 }
