+++ core/src/main/java/org/kiwix/kiwixmobile/core/main/CoreMainActivity.kt
@@ -75,6 +75,7 @@ abstract class CoreMainActivity : BaseActivity(), WebViewProvider {
   abstract val historyFragmentResId: Int
   abstract val notesFragmentResId: Int
   abstract val helpFragmentResId: Int
+  abstract val zimHostFragmentResId: Int
   abstract val cachedComponent: CoreActivityComponent
   abstract val topLevelDestinations: Set<Int>
   abstract val navHostContainer: FragmentContainerView
@@ -203,9 +204,15 @@ abstract class CoreMainActivity : BaseActivity(), WebViewProvider {
       R.id.menu_notes -> openNotes()
       R.id.menu_history -> openHistory()
       R.id.menu_bookmarks_list -> openBookmarks()
+      R.id.menu_host_books -> openZimHostFragment()
       else -> return false
     }
     return true
+  }
+
+  private fun openZimHostFragment() {
+    navigate(zimHostFragmentResId)
+    handleDrawerOnNavigation()
   }
 
   private fun openHelpFragment() {
