+++ core/src/main/java/org/kiwix/kiwixmobile/core/webserver/WebServerHelper.kt
@@ -1,6 +1,6 @@
 /*
  * Kiwix Android
- * Copyright (c) 2019 Kiwix <android.kiwix.org>
+ * Copyright (c) 2023 Kiwix <android.kiwix.org>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -15,7 +15,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package org.kiwix.kiwixmobile.webserver
+package org.kiwix.kiwixmobile.core.webserver
 
 import android.util.Log
 import io.reactivex.Flowable
@@ -26,7 +26,7 @@ import org.kiwix.kiwixmobile.core.utils.ServerUtils
 import org.kiwix.kiwixmobile.core.utils.ServerUtils.INVALID_IP
 import org.kiwix.kiwixmobile.core.utils.ServerUtils.getIp
 import org.kiwix.kiwixmobile.core.utils.ServerUtils.getIpAddress
-import org.kiwix.kiwixmobile.webserver.wifi_hotspot.IpAddressCallbacks
+import org.kiwix.kiwixmobile.core.webserver.wifi_hotspot.IpAddressCallbacks
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
