+++ core/src/main/java/org/kiwix/kiwixmobile/core/webserver/ZimHostCallbacks.kt
@@ -1,6 +1,6 @@
 /*
  * Kiwix Android
- * Copyright (c) 2020 Kiwix <android.kiwix.org>
+ * Copyright (c) 2019 Kiwix <android.kiwix.org>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -15,17 +15,12 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package org.kiwix.kiwixmobile.webserver
+package org.kiwix.kiwixmobile.core.webserver
 
-import org.kiwix.kiwixmobile.core.base.BaseContract
-import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem
-
-class ZimHostContract {
-  interface View : BaseContract.View<Presenter> {
-    fun addBooks(books: List<BooksOnDiskListItem>)
-  }
-
-  interface Presenter : BaseContract.Presenter<View> {
-    fun loadBooks(previouslyHostedBooks: Set<String>)
-  }
+interface ZimHostCallbacks {
+  fun onServerStarted(ip: String)
+  fun onServerStopped()
+  fun onServerFailedToStart()
+  fun onIpAddressValid()
+  fun onIpAddressInvalid()
 }
