+++ core/src/main/java/org/kiwix/kiwixmobile/core/webserver/ZimHostPresenter.kt
@@ -1,6 +1,6 @@
 /*
  * Kiwix Android
- * Copyright (c) 2019 Kiwix <android.kiwix.org>
+ * Copyright (c) 2023 Kiwix <android.kiwix.org>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -15,7 +15,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package org.kiwix.kiwixmobile.webserver
+package org.kiwix.kiwixmobile.core.webserver
 
 import android.util.Log
 import io.reactivex.SingleObserver
@@ -24,8 +24,8 @@ import org.kiwix.kiwixmobile.core.base.BasePresenter
 import org.kiwix.kiwixmobile.core.data.DataSource
 import org.kiwix.kiwixmobile.core.di.ActivityScope
 import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem
-import org.kiwix.kiwixmobile.webserver.ZimHostContract.Presenter
-import org.kiwix.kiwixmobile.webserver.ZimHostContract.View
+import org.kiwix.kiwixmobile.core.webserver.ZimHostContract.Presenter
+import org.kiwix.kiwixmobile.core.webserver.ZimHostContract.View
 import javax.inject.Inject
 
 @ActivityScope
