+++ core/src/main/res/drawable/ic_share_35dp.xml
