+++ core/src/main/res/layout/item_custom_spinner.xml
